+++ CHANGELOG.md
@@ -6,3 +6,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 ## 3.4.1.0
 
 - Add progress dialog when checking for Curse mod updates [#409]
+- Check all selected Curse mods for updates from Edit Mods [#409]